Separate out some underlying libp2p components.
