+++ rust/noosphere-name-system/src/dht/mod.rs
@@ -1,8 +1,10 @@
+mod behaviour;
 mod channel;
 mod client;
+mod node;
 mod swarm;
 mod types;
 
 pub use client::DHTClient;
-pub use swarm::DHTSwarm;
+pub use node::DHTNode;
 pub use types::DHTConfig;
