+++ rust/noosphere-name-system/src/dht/types.rs
@@ -2,13 +2,15 @@ use crate::dht::channel::{Message, MessageClient, MessageProcessor};
 use libp2p;
 
 pub enum DHTRequest {
-    GetRecord(Vec<u8>),
+    GetRecord { name: Vec<u8> },
     SetRecord { name: Vec<u8>, value: Vec<u8> },
+    StartProviding { name: Vec<u8> },
 }
 
 pub enum DHTResponse {
     GetRecord { name: Vec<u8>, value: Vec<u8> },
     SetRecord { name: Vec<u8> },
+    StartProviding { name: Vec<u8> },
 }
 
 pub type DHTMessage = Message<DHTRequest, DHTResponse>;
